pyproject.toml
[tool.poetry]
name = "ant_project"
version = "0.1.0"
description = ""
authors = ["c95ths <ford19960516@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tensorflow-gpu = "2.4.0"
scipy = "^1.6.0"
matplotlib = "^3.3.3"
mne = "^0.22.0"
pandas = "^1.2.0"
PyWavelets = "^1.1.1"
tqdm = "^4.59.0"
sklearn = "^0.0"
umap-learn = "0.4.0"
seaborn = "^0.11.1"
datashader = "^0.12.1"
bokeh = "^2.3.1"
holoviews = "^1.14.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
